# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API for Driving License Management",
        "title": "Driving License REST API",
        "contact": {
            "url": "https://github.com/adohong4"
        },
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/licenses/create": {
            "post": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "Create a new driving license entry",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "DrivingLicense"
                ],
                "summary": "Create a new driving license",
                "parameters": [
                    {
                        "description": "Driving License object",
                        "name": "driving_license",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.DrivingLicense"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.DrivingLicense"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httpErrors.RestError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httpErrors.RestError"
                        }
                    }
                }
            }
        },
        "/licenses/getAll": {
            "get": {
                "description": "Get a paginated list of all active driving licenses",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "DrivingLicense"
                ],
                "summary": "Get all driving licenses",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page size",
                        "name": "size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.DrivingLicenseList"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httpErrors.RestError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httpErrors.RestError"
                        }
                    }
                }
            }
        },
        "/licenses/search": {
            "get": {
                "description": "Search for driving licenses by license number with pagination",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "DrivingLicense"
                ],
                "summary": "Search driving licenses by license number",
                "parameters": [
                    {
                        "type": "string",
                        "description": "License number to search",
                        "name": "license_no",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page size",
                        "name": "size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.DrivingLicenseList"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httpErrors.RestError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httpErrors.RestError"
                        }
                    }
                }
            }
        },
        "/licenses/{address}": {
            "get": {
                "description": "Get a single driving license by its Wallet Address",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "DrivingLicense"
                ],
                "summary": "Get driving license by Wallet Address",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Driving License Wallet Address",
                        "name": "address",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.DrivingLicense"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httpErrors.RestError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/httpErrors.RestError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httpErrors.RestError"
                        }
                    }
                }
            }
        },
        "/licenses/{id}": {
            "get": {
                "description": "Get a single driving license by its ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "DrivingLicense"
                ],
                "summary": "Get driving license by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Driving License ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.DrivingLicense"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httpErrors.RestError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/httpErrors.RestError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httpErrors.RestError"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "Update an existing driving license by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "DrivingLicense"
                ],
                "summary": "Update a driving license",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Driving License ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Driving License object",
                        "name": "driving_license",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.DrivingLicense"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.DrivingLicense"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httpErrors.RestError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httpErrors.RestError"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "Soft delete a driving license by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "DrivingLicense"
                ],
                "summary": "Delete a driving license",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Driving License ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Driving License object (for modifier)",
                        "name": "driving_license",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.DrivingLicense"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.DrivingLicense"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httpErrors.RestError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httpErrors.RestError"
                        }
                    }
                }
            }
        },
        "/licenses/{id}/add-wallet": {
            "put": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "Add or update wallet address for a driving license",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "DrivingLicense"
                ],
                "summary": "Add wallet address",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Driving License ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Wallet address",
                        "name": "wallet",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.DrivingLicense"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.DrivingLicense"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httpErrors.RestError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httpErrors.RestError"
                        }
                    }
                }
            }
        },
        "/licenses/{id}/confirm-blockchain": {
            "put": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "Update blockchain transaction hash and set on_blockchain to true",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "DrivingLicense"
                ],
                "summary": "Confirm blockchain storage",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Driving License ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Blockchain details",
                        "name": "blockchain",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.DrivingLicense"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.DrivingLicense"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httpErrors.RestError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httpErrors.RestError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "httpErrors.RestError": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
        "models.DrivingLicense": {
            "type": "object",
            "required": [
                "id"
            ],
            "properties": {
                "active": {
                    "type": "boolean"
                },
                "authority_id": {
                    "description": "Mã nơi cấp",
                    "type": "string"
                },
                "avatar": {
                    "type": "string"
                },
                "blockchain_txhash": {
                    "description": "Mã lưu ở blockchain",
                    "type": "string"
                },
                "created_at": {
                    "description": "Thời gian tạo",
                    "type": "string"
                },
                "creator_id": {
                    "description": "ID của người tạo",
                    "type": "string"
                },
                "dob": {
                    "description": "Ngày sinh",
                    "type": "string"
                },
                "expiry_date": {
                    "description": "Ngày hết hạn (có thời hạn, vô thời hạn)",
                    "type": "string"
                },
                "full_name": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "identity_no": {
                    "description": "Căn cước công dân",
                    "type": "string"
                },
                "issue_date": {
                    "description": "Ngày cấp",
                    "type": "string"
                },
                "issuing_authority": {
                    "description": "Nơi cấp",
                    "type": "string"
                },
                "license_no": {
                    "description": "Số bằng lái",
                    "type": "string"
                },
                "license_type": {
                    "description": "Loại bằng lái (A1, B1, B2, ...)",
                    "type": "string"
                },
                "modifier_id": {
                    "description": "ID của người sửa",
                    "type": "string"
                },
                "nationality": {
                    "description": "Quốc tịch (Việt Nam, Hàn Quốc, ....)",
                    "type": "string"
                },
                "on_blockchain": {
                    "description": "Trạng thái lưu ở blockchain (lưa/ chưa lưu)",
                    "type": "boolean"
                },
                "owner_address": {
                    "description": "Địa chỉ",
                    "type": "string"
                },
                "point": {
                    "description": "Điểm bằng lái xe (0 \u003c point \u003c 12)",
                    "type": "integer"
                },
                "status": {
                    "description": "Trạng thái (pending: chờ đợi, expiry: hết hạn, active: hoạt động, pause: tạm dừng (point = 0))",
                    "type": "string"
                },
                "updated_at": {
                    "description": "Thời gian cập nhật",
                    "type": "string"
                },
                "version": {
                    "description": "Phiên bản, tự động tăng",
                    "type": "integer"
                },
                "wallet_address": {
                    "type": "string"
                }
            }
        },
        "models.DrivingLicenseList": {
            "type": "object",
            "properties": {
                "driver_licenses": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.DrivingLicense"
                    }
                },
                "has_more": {
                    "type": "boolean"
                },
                "page": {
                    "type": "integer"
                },
                "size": {
                    "type": "integer"
                },
                "total_count": {
                    "type": "integer"
                },
                "total_pages": {
                    "type": "integer"
                }
            }
        },
...
    }
}